requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1

# LangChain + integrations
langchain==0.2.6
langchain-community==0.2.6
langchain-groq==0.1.3
langchain-chroma==0.1.2

# Pin Chroma to a stable version
chromadb==0.4.22

# HuggingFace + loaders
huggingface_hub==0.23.4
unstructured==0.15.0